Add basic test parallelization
Adds basic parallelization of testing. Since backends may or may
not implement a mechanism for this we use the most generic approach
and parallelize the number of tests being run that belong to a
specific distribution. This means running the tests for a single
distribution will not see any differences even if it has multiple
test files.

This is still buggy when using a degree > 1 -- tests can fail
with compile time errors for not finding dependencies in a seemingly
random fashion.